# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Linkedofy",
        "description": ""
    },
    "tags": [
        {
            "name": "Linkedofy",
            "description": ""
        }
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": 
    {
        "/withdraw_invite_post": 
        {
            "delete": {
                "tags": [
                    "Delete"
                ],
                "summary": "",
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": 
                    {
                        "application/json": 
                        {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
        "/autoconnect": 
        {
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "",
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": 
                    {
                        "application/json": 
                        {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
        "/broadcast": 
        {
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "",
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": 
                    {
                        "application/json": 
                        {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
        "/broadcast/{license}": 
        {
            "get": {
                "tags": [
                    "Get"
                ],
                "summary": "",
                "parameters": [
                    {
                        "in": "path",
                        "name": "license",
                        "required": true,
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/license"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "not found."
                    }
                }
            }
        },
        "/autoconnect/{license}": 
        {
            "get": {
                "tags": [
                    "Get"
                ],
                "summary": "",
                "parameters": [
                    {
                        "in": "path",
                        "name": "license",
                        "required": true,
                        "description": "user with id",
                        "schema": {
                            "$ref": "#/definitions/license"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "not found."
                    }
                }
            }
        }
    },
    "definitions": {
        "license": {
            "properties": {
                    "type": "string"
            }
        },
        "user_info": {
            "type": "object",
            "properties": {
                "license": {
                    "type": "string"
                },
                "profile_url": {
                    "type": "string"
                }
            }
        },
        "user": {
            "type": "object",
            "properties": {
                "users": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/user_info"
                    }
                }
            }
        }
    }
}